update status for rider kyc

+++ wallet/serializers.py
@@ -40,7 +40,7 @@ class GetTransactionsSerializer(serializers.ModelSerializer):
 
 
 class AddCardSerializer(serializers.Serializer):
-    amount = serializers.FloatField()
+    amount = serializers.DecimalField(max_digits=10, decimal_places=2)
 
 
 class ChargeCardSerializer(serializers.Serializer):
